Fix bug that prevents test results/detail/history from displaying.
If a project fails to build, an alert box will pop up saying so if one
tries to view test history or detail. A future change will need to state in
the table that a package was not build and thus does not have test results.
It should then also disable the select button or something similar.

Also, there needs to be a change where the hardcoded filenames have global
values instead.

Change-Id: I365bf0e9b9f5c5fd97cb9d9d87c7eea4cbc91026