Add bootstrap-table addon. Add pagination and column sorting to batch list/select table.
Added bootstrap-table addon that makes working with tables much simpler, and will
allow us to implement additional features that would normally be much more difficult.
There was no simple way to do pagination and sorting with what we had before,
aside from generating the html within javascript.

I removed the actions column. Each piece of data in a table should be unique to that specific
element. The actions are the same for each element in the table, so they should be outside the table.
This is much cleaner in my opinion and really the only option with this framework addon.
Also if we ever decide to add more and more actions, that column will get out of control and distracts
the user from what they actually care about, the package. We can discuss this further.

Change-Id: I4726268c79b1c3ef7b2e5292c75ae9c09ae5f14d