Manage Single Server Screen:
Changed the build server drop-down to use bootstrap's multi-select plug-in
The options are listed as radio buttons with none of the build servers selected by default.

Changed the build server drop down to display labels instead of names.

Also deleted the json file fetched from the Jenkins master and its containing folder once a read is done on it

Change-Id: I48ffd5ea0051a4f7e89bbf8db80687140c4ba4ef