Provide server side managed package info for listManagedPackage()
This patch shows how the managed package list is parsed to provide
the managed package version.  It manages minor O/S version differences
such as Ubuntu 14.04.1 and Ubuntu 14.04.2 without having to record each
update in the managed list.   We don't want to update the managed list
every time a minor release is made generally available.

getManagedPackage() checks both the os release and version fields of
the specified node to find a match.  Since only the definition for
Ubuntu 14.04, it finds that for both 14.04.1 and 14.04.2

Note the data structure would allow us to add a definition for 14.04.1
ahead of 14.04 if we wanted to define a specific managed versions but we
only need to do that if we are going to override O/S defaults.

To Do -

1) UI
2) Add a column for O/S version
3) Make the code work when no package is specified.  I limited it to
   conform with the existing code.
4) Add bootstrap tables

Change-Id: I5ef2fa13e67ac4096365ee5ca6afb84b860678ae