Code Changes for uploading rpm,deb and tar packages to custom repository and uploading chef data during Application initialization.
The code changes are done to handle following scenarios:

a) Transferring chef data to Jenkins Master for the autoport host instance and performing cookbook upload operation using knife.

- When we start autoport instance , the complete chef-repo which is part of autoport code repository , is compressed into a tar file and copied to /data/config directory of the local autoport instance.  Based on the comparison of cookbook version maintained in ManagedList.json  , tar file is uploaded to shared location on Jenkins master. 

Using remote execution of command on Jenkins Master,chef-repo.tar is extracted and all of the required chef-data is uploaded to chef-server using knife command. (Note , we are not using jenkins job to carry out this step , rather making use of python Paramiko module to carry out remote execution).

Current code drop has all the required configuration files and the keys pushed into the chef-repo (.chef) , such that knife commands can be directly executed without any further configuration, once the chef-tar is extracted in the shared location of the Jenkins Master.

b) Uploading packages (rpm,debs,tar) to the custom repository.

-/data/packages is the directory created to temporarily hold the uploaded packages , till they are ftp'd to the custom repository on Jenkins Master.

Path where the package is to be placed on Jenkins Master is decided based on extension of the package. We do not maintain a local package directory tree (/data/packages, /data/packages/rpm , /data/packages/debs , /data/packages/tar), as  the repository is already pre-created over the Jenkins Master with required directory structure and other configuration files to allow them to act as yum and apt repository.

Change-Id: If8491618159cec57964ba911e933b42fd1137710