Continuation of Caroline's patch -- Caroline added support for user based managed list.  /var/opt/autoport/user/<hostname>/ManagedList.json -- Luke added Chef support.  It is located in the same fixed location and is shared by all users.    Decoupled chef versioning from the ManagedList by adding a separate file for that purpose. /var/opt/autoport/chef-cookboot-version.json
The reason for this patch is that there is user based data in the managed list meaning that it needs to be
recorded in a user specific  location and hence decoupled from chef support as that is shared.  Build slaves
expect to find chef data at a fixed location on Jenkins master.

I reset the sequence numbers to 1 as the location of the ManagedList.json file is new.

Change-Id: I68d9067d093ba087a2c7c7a5a235dcb615cd153b