Fix parsing of environment variables in readme files within build analyzer
Hadoop yields an invalid environment variable.  "MAVEN_OPTS."  The expected
string is MAVEN_OPTS="-Xms256m -Xmx512m".  It was locating the wrong sentence
in the readme file because the search term was MAVEN_OPTS instead of
MAVEN_OPTS=.  The second problem was that the expected value is a quoted string.

Change-Id: I4980fd9fdb93d9ff7bdf36020ca7622921e64acf