Updated code to copy .chef folder to jenkins home directory on autoport initialization and removed chef-server certificates from autoport code
The code changes are done with following approach in mind. 

1.The .pem files (autoport-ibm-validation.pem and autoport-chef.pem) will not be part of autoport code. This is to address the issue where .pem file in autoport-code is over-writing the new .pem files created during a fresh chef-server setup. 

a) In case of cloud context docker base image or docker-file should be pre-canned with the required .pem files in the directory /var/opt/autoport/chef-repo/.chef. (chef-setup done once with all configuration such as user and org preserved in docker-file) 

b) In case of non-cloud infrastructure user performing the chef-server setup would be responsible to place the keys at the required location (/var/opt/autoport/chef-repo/.chef). 

At the time of initialization autoport driver would copy the chef-repo.tar as usual. On extraction of tar file we would get chef-related stuff except for the keys which are already present based on point a) and b). 

2.On initialization apart from copying chef-repo to shared location (/var/ww/autoport), code is updated to copy the .chef folder to jenkins home directory, automating one of the steps in wiki related to (chef-workstation setup). 

Change-Id: I421010e384c308420ff557421e0386516554a1fd