Add packages for RHEL 7.1 to Managed Runtime
Note we shouldn't be specifying version numbers for python as
python2 is the default.  If you perform a package search on "python"
you get the latest version 2 which is what we want.  RHEL or Ubuntu
may provide a bug fix to python2 which causes the version to change
in a minor way. This in fact happened on RHEL.  We had the wrong
version listed in the Managed List.

Also, we want our definition for each release to last more than a
service update.  Versions are updated with each service release,
ie. Ubuntu 14.04.1 to Ubuntu 14.04.2.  We really shouldn't specify
versions for any package that is provided by the operating system,
but never say never:-)

Change-Id: I517d693b6f4ba218324b4052c64a530a80fbf404