Improve performance with moveArtifacts by eliminating jenkins connection for each invocation.
This is implemented by caching the ssh/ftp points on each thread in the threadPool which necessitates
the use of thread local storage so that each thread has a private connection.  All invocations to
moverArtifacts needed to be made from the thread pool.

There were two cases where moveArtifacts was being invoked under the context of the process in a
synchronous fashion.  These were switched to use the thread pool which necessitated the use of thread
condition variables to synchronize the caller and moveArtifacts

A new module is introduced to provide the jenkins master mover function called mover.py.  A connections
to jenkins master are established at startup.  The rest are connected on the fly.  Retry logic is included
so that authentication issues are automatically managed if a connection expires.

Increased the size of the thread pool to 16 to accomodate a total of 8 build servers.  There is a
2-1 thread to build server ratio as the list managed runtime function uses 2 thread per build server.  One
directly and one indirectly as a routine that is called down the stack invokes moveArtifact.

We connect 3 threads to jenkins master at startup to keep the startup time down.

Change-Id: I0a069d0d76c3ec5fd8e47bc6cbdb98c7df844b78