Meta.arti file is almost never produced which breaks the report code.
Change the way the jobs are built and tested, so that the meta.arti
file is always produced.

Modified config template so that build, test, and install commands
are supported.  Later, we will use the install command for packaging
and also potentially to allow projects to be installed as a function
of processing a batch file. This is desirable for resolving
prerequisites.

Fixed the report code so that it references PrimaryLanguage, not the
build system which is really a command, but sometimes says a lang.

Change-Id: I23b8a6fd788060e648c34d1b494219d73e3c1f84