Improved environment variable generation in the buildAnalyzer.  It was picking up CFLAGS="-m32" for redis which caused a build failure as 32-bit applications cannot be compiled on the latest versions of Ubuntu and RHEL.
Added an array of substrings that cannot be included in a match.  It contains
"32" to eliminate 32-bit apps without the "-m" so that it can handle "-p32".
The occurence of 32 is most likely related to 32-bit apps than all other cases.

Added some strings to avoid platform specific environment variables as our
build commands are applied across all build slaves.  Command generation has to
be platform independent.  ie. added arm, ppc, x86

We can add additional substrings to this array in the future to address specific
build and test failures

Also prepended "./" in front of shell scripts provided by the project.   We were
seeing an error with the grep substitution with the project "the_silver_searcher"
where the command was documented in the README as "build.sh".  This results in a
command not found error if we invoke the string as-is.

Fixed the suffix for grepped commands to redirect output to the correct *.arti file.

Change-Id: Ia8bfe261b982d767c3806871adcf2698dd325acb