Fixed an availability issue with client connection hanging.  This occurs when a batch job is running as this usually means one of the build servers is busy.  The node query jenkins job was put at the back of the queue.
The solution is to return cached node information for client
connections.  The jenkins job query node is not needed.  Client
connections now take seconds instead of minutes.

Also fixed changing jenkins master.

This also fixes a usability problem where the user re-loads the
page.  When this happens, the user gets the initial page, but the
state of the backend is not impacted.

A PROBLEM TO BE FIXED:

The "Select build server" drop down menu still shows old build servers
after changing the jenkins master in the settings menu.  The javascript
code has been updated with the new server information, but it appears to
be the case that the bootstrap library is caching the old data.  How do
you tell bootstrap to re-load.  If you reload the web-page, you get the
correct data.

Change-Id: I00c91be5c29fa01b754df1082ded41dda981f04f