Changes done to fetch ipaddress of each slave node and carrying out knife boostrap using ipaddress rather than hostname
As per the previous comment manually removed the chef-repo-version.json
on soe-test1 and soe07x-vm2 and refreshed by starting autoport driver.

Each of the above node is also reconfigured with chef-server to work with ipaddress rather than fqdn (as per steps in wiki).

Once the code is merged we will send out a note to the team to remove chef-repo-version.json if they are using jenkins-master other than soe-test1 or soe07-vm2
and then restart autoport driver with newly pulled in code.

Change-Id: I91cdd07ff4dc4e3dae2be491b946a766dbc8c147