Place autoport commands and build results in a scratch directory, so as not to interfere with the build and test of a project.
We were seeing errors with memcached.

The packaging code will need a lot of scratch space as an rpm
needs to be unrolled.  This isolates that code from the project.

The .gitignore file is placed in the .autoport_scratch/ so that
testcases that use github tools will ignore it.   memcached does.

Change-Id: I68cd62b99131fb420473520abada6e0c4cfae0d8