Complete travis.yaml code changes.
Added support for environment variables.
Added support for before_install and install sections as these sections are preliminary
to build which almost never works without these steps.  Travis.yaml is a large percentage
(>25%) of C/C++ programs and is therefore needed to improve our overall build and test success.

The only down side is the install of packages which is easily dealt with in the cloud
by restoring the snapshot of the build server.  This is what Travis CI does.

The remaining work item is to test on RHEL.  Most of the install commands seem to be
based on ubuntu. Presumably these commands fail which does not impact the over all
execution of the phase.  Commands are separated with a ';'.

Change-Id: Ic2a903dd79749acbba6a9a4e507021988b000b93