Added guard conditions and enabled ignore_failure flag for most of the resources.
Had to add ignore_failure flag for each resource as there was no other way to set this flag at recipe level or skip the rest of the recipe whenever a failure occurs. Guard conditions were added so that failure of a previous resource which was ignored due to flag being true should not lead to direct execution of next resource which could leave things in inconsistent state.

Changes are done to ignore failure at least in below 2 cases:
a) When any operation related to apt/yum package fails 
b) When a particular archive/binary is not found in repository. 

Change-Id: I202ff2f7471b9c0aab1bdbfb58396ae8133e0841