ZChanged batch file processing so that one thread is used to move artifacts for the whole file, instead of potentially the entire pool of threads.  This will help the multi-user aspects of the tool tremendously.
Also fixed the 'gsa' bug when removing local batch test results.

Changed a couple of warning messages to debug, since it is not
an issue for build results for a given project to be missing.  There
is no requirement that every job that is started produce build
artifacts.  An individual job may fail for any number of reasons.

Change-Id: I40f4bdee778435387ffcb75bf0fa0d8f55f4b17f