Initial code drop for CentOS
Node comes up and is recognized.  Build Server drop-down button identifies CentOS node correctly.

Having a problem building junit on CentOS.  I believe this is due to the firewall.  Please investigate
getting ftp on the node so that we can authenticate with BSO.

Added build slave to Jenkins Developer Cluster in Austin.

Change-Id: Ib9e639fcb84f3ee070a8fcef790b7b2422cc7b04