This improves web performance of Jenkins dashboard. (ie. http://soe07x-vm2.aus.stglabs.ibm.com:8080/)
This is a helper script to manually clean all but the 2000 most
recent jenkins jobs on jenkins master.  You need to transfer this
command to jenkins master and invoke it.  It asks you to confirm
that you want to remove files and is therefore safe to invoke.

After removing older files, you will need to manually re-load
the disk configuration on jenkins master using the Jenkins
dashboard.  The jenkins dashboard should perform well now.

This is not intended for end users now, but may be integrated into
the autoport UI in the future.  The existing garbage collection of
jenkins master that is embodied in Jenkins plugin to discard old
jobs does not work as it does not remove the last successful job.

Here I am talking about:

  <logRotator class="hudson.tasks.LogRotator">
    <daysToKeep>7</daysToKeep>
    <numToKeep>3000</numToKeep>
    <artifactDaysToKeep>7</artifactDaysToKeep>
    <artifactNumToKeep>3000</artifactNumToKeep>
  </logRotator>

Change-Id: Id8d5ab271bf8e0fa85d1d270dd0b150ac0bc36b0