Changing threadPool size.
Its been observed that if we start building a single project 
on all the 4 build-servers, 4 threads out of 8 would be consumed from threadPool.At this time if we trigger "List All" operation on build tab we would not be able to fetch the result immediately as we require a total of 8 threads, but we are left with only 4 threads. This would block list operation till the build on the nodes finish. Hence increasing the threadPool size to overcome this problem temporarily.

Change-Id: I463b46aceed8211c574f0455fc6bcdcee3384516