Add support for cloud deployments where a single floating IP is used for the entire cluster.  A reverse proxy, NGINX, is used to map the URL to either autoport driver or Jenkins master which exist on different servers.  Autoport internally connects with the jenkins master via sftp via an internal IP ADDR and it presents a URL based interface to Jenkins to users.  Autoport therefore needs to know both the internal IP Address and external URL of Jenkins master.  There is an assumption that for the cloud the internal IP Address is static.  That is, all autoport services configure a private jenkins master at the same IP address.  Autoport doesn'tassume this though.  It defines a command line interface to specify both.   There is also a config.ini that has been the internal hostname (or ip address) and the external jenkins url.
This patch provides the base enablement for cloud installations.  The cloud
deployment script must be updated to either set the config.ini field or specify
the command line argument.

These are the relevant fields in the config.ini file.

jenkinsUrl: http://soe-test1.aus.stglabs.ibm.com:8080
jenkinsHostname: <host or ipaddr or empty>

Note the field above does not have to be modified for standalone clusters.  The
hostname is deduced from the URL if the value above contains a "<" or is empty.

This is the command line argument:

python main.py -j -j 9.3.126.17

The jenkins URL is no longer in the settings menu.

Change-Id: I623ead567a993b4ce04ebe5690983e20cf10ac6b