Moved environment file for advance tool chain from files folder to template folder to overcome issue of differences in line breaks in DOS and UNIX.
While updating supervessel nodes, it was observed that chef cookbook files had dos format(may be autoport code on driver was transferred from Windows machine). 
Because of this when at.sh file was deployed on the salve node during sync it led to a wrong format and I had to correct it using dos2unix utility.

To overcome this I have now moved at.sh from cookbook's files/default folder to
template/default folder. Chef-client would take care of the format while deploying the template on the target node.


Change-Id: I93f0d42de621abf50c0b442d2ed8512e49fed23d